requirements.txt
asgiref==3.7.2
black==23.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudinary==1.36.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.57.0
django-braces==1.15.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.1
django-model-utils==4.3.1
django-star-ratings==0.9.2
django-starfield==1.0.post2
gunicorn==21.2.0
idna==3.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
urllib3==2.0.6
